+++ app/src/main/java/com/example/planet/data/remote/dto/request/plogging/TrashImageUrlInfo.kt
@@ -2,7 +2,7 @@ package com.example.planet.data.remote.dto.request.plogging
 
 data class TrashImageUrlInfo(
     val userId: Long,
-    val ploggingId: Int,
+    val ploggingId: Long,
     val imageUrl: String,
     val longitude: Double,
     val latitude: Double,
