+++ app/src/main/java/com/example/planet/data/remote/dto/response/plogging/PloggingComplete.kt
@@ -1,6 +1,6 @@
 package com.example.planet.data.remote.dto.response.plogging
 
 data class PloggingComplete(
-    val ploggingId: Int,
+    val ploggingId: Long,
     val message: String
 )
