+++ app/src/main/java/com/example/planet/data/remote/dto/response/plogging/PloggingId.kt
@@ -1,7 +1,7 @@
 package com.example.planet.data.remote.dto.response.plogging
 
 data class PloggingId(
-    val ploggingId: Int,
+    val ploggingId: Long,
     val userId: Long,
     val weight: Double,
     val height: Double,
