+++ app/src/main/java/com/example/planet/domain/usecase/plogging/GetPloggingInfoUseCase.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class GetPloggingInfoUseCase @Inject constructor(private val ploggingRepository: PloggingRepository) {
-    suspend operator fun invoke(ploggingId: Int): Flow<ApiState> {
+    suspend operator fun invoke(ploggingId: Long): Flow<ApiState> {
         return ploggingRepository.getPloggingInfo(ploggingId)
     }
 }
